#337 Tweaks. Move webhelp to core.
Webhelp can be launched in CLI/Console mode. Is it's own plugin. The previous webhelp calls use the new methods.
